patch.js: Add min_build (from {{thcrap_patch_info}}) and global server description URL
After considering this for a long time, this ended up being the preferred way to check for new versions of the engine itself. Basically, each patch can define a minimum version it recommends to run (I'm not saying 'it requires' here, as right now, everything is still backwards-compatible), and the engine can then display a nag screen, prompting to user to re-download the archive.

This has the advantage that, just as anywhere else, no centralized source of the patch needs to be stated anywhere. If some group uses a modified version of thcrap for some patch, they can direct users to their own site (using desc_url, the "main descriptive page" of a server).