TPCFmtSpells.php: Don't escape spell cards after all.
We needed to do this because spell names are actually sprintf format strings, even though the games never actually take advantage of this. Thus, it should be our goal to *remove* these unnecessary sprintf calls *in Touhou code* and not to work around them, as they have been always vulnerable to buffer overflows.