Remove TPCFormat.
Listen. I know that MediaWiki is the completely wrong tool for our job, so. This was the most ridiculous and nonsensical out of all the parsers here - each new format would have required yet another adaption of the code... and the corresponding templates, if we ever had one.

And all that just so that we can get the JSON result we want. Just upload the file as you need it, and it's not like this needs to be stacked with anything on the wiki's end. Problem solved.