Remove {{lang}} wrapper templates.
I've removed them from the pages on Touhou Patch Center as well, this is just to make sure we'll never see them again.

It's an ugly and confusing wart taken over from Touhou Wiki that was only ever intended for display issues with CJK characters. Unfortunately, it soon spread to any other case of mixed languages in one string, since it made everything look nicer with the commonly used fonts. Ultimately though, this is the fault of font designers not designing balanced type faces with a pleasing appearance for any combination of scripts.

Even if we did need custom font parameters for scripts one day, the text parser should automatically detect and set these based on the Unicode code points of the text content itself, not manual markup.