Add a base class for templates rendered by this extension.
Reducing duplication where we can. I'm a bit reluctant to get rid of the MagicWordwgVariableIDs hook, though - enumerating subclasses of a class in PHP requires getting a list of *all classes declared so far* and then checking the relationship of each of them.

As a result, I don't really see any way to automate setting the $magicWords variable from the i18n file either.