Rename the patch list "server" concept to "repository".
Prior to this, we had two things called "server": the list of patches downloadable from one place (stored in server.js) and the actual download URLs (stored in the "servers" array of server.js and patch.js). This has been getting really confusing and kind of limits the function naming in the client side code.