MagicWordwgVariableIDs → GetMagicVariableIDs
Another hook deprecated in MediaWiki 1.35. The new extension.json
system really prefers those to be at extension registration level as
well. I really don't like the redundancy of having to both autoload the
class and explicitly add its hook in the same place… but conformance
seems better in this regard.