TouhouPatchCenter: Rework registration of patch generation hooks.
This one, however, is going to stay as it was… for the most part.
After all, these aren't really the MediaWiki kind of "hooks" that are
called on specific events, so it's not worth it to turn them into
something else. *Especially* since we're now supposed to use that new
interface-based hook system, introduced in MediaWiki 1.35:

	https://www.mediawiki.org/wiki/Manual:Hooks#Handling_hooks_in_MediaWiki_1.35_and_later

(Which is the third deprecation in 1.35 we've encountered so far. Truly
one massive release… good that we've waited so long with the update.)

So, no autoloading, and still explicit old-style registration at file
scope. Except that I made the mistake of forgetting the `global`
declaration during the first attempt, which led me to throw away these
globals althogether. Much cleaner this way!