Content::getNativeData() → TextContent::getText()
Deprecated in MediaWiki 1.33, because it hardcoded the assumption of
the default wikitext content model:

	https://phabricator.wikimedia.org/T155582

So, extensions have to enforce the content model themselves, really.
We've even been doing this in one place already, as of 3566677.

But wow, how did we even end up with WikiPage::getText() in b82bf2d?!?
That function already was deprecated in MediaWiki 1.21, which had
already been released by the time that commit was written. That one was
removed way back in 1.29.