TouhouPatchCenter: React to file undeletion.
We could optimize this and only run evalFile() if the entire page is
restored, because undeleting old versions doesn't change the active
version of a file. But relying on $fileVersions being empty in case of
a full page undeletion seems rather daring, especially since this array
just contains revision IDs. Let's keep it simple.